Move MemoryPool into its own file. Add a resize method, so that rented buffers are returned before resizing. Ensure BinaryUtil ustilises new byte arrays and doesn't double-rent buffers with no returns.
