+++ src/Elasticsearch.Net/Utf8Json/Formatters/DateTimeFormatter.cs
@@ -929,7 +929,7 @@ namespace Elasticsearch.Net.Utf8Json.Formatters
             var day = 0;
             if (hasDay)
             {
-                var poolArray = JsonSerializer.MemoryPool.Rent();
+                var poolArray = MemoryPool.Rent();
                 try
                 {
                     for (; array[i] != '.'; i++)
@@ -941,7 +941,7 @@ namespace Elasticsearch.Net.Utf8Json.Formatters
                 }
                 finally
                 {
-					JsonSerializer.MemoryPool.Return(poolArray);
+					MemoryPool.Return(poolArray);
                 }
             }
